fix Ping
