--- app/src/main/java/de/fraunhofer/fokus/OpenMobileNetworkToolkit/DataProvider/PingInformation.java
@@ -1,18 +1,0 @@
-/*
- *  SPDX-FileCopyrightText: 2023 Peter Hasse <peter.hasse@fokus.fraunhofer.de>
- *  SPDX-FileCopyrightText: 2023 Johann Hackler <johann.hackler@fokus.fraunhofer.de>
- *  SPDX-FileCopyrightText: 2023 Fraunhofer FOKUS
- *
- *  SPDX-License-Identifier: BSD-3-Clause-Clear
- */
-
-package de.fraunhofer.fokus.OpenMobileNetworkToolkit.DataProvider;
-
-public class PingInformation {
-    private Double unixTimestamp;
-    private int icmpSeq;
-    private int ttl;
-    private double rtt;
-
-    private String host;
-}
