--- app/src/main/java/de/fraunhofer/fokus/OpenMobileNetworkToolkit/Iperf3/METRIC_TYPE.java
@@ -1,8 +1,0 @@
-package de.fraunhofer.fokus.OpenMobileNetworkToolkit.Iperf3;
-
-public enum METRIC_TYPE {
-        THROUGHPUT,
-        RTT,
-        JITTER,
-        PACKET_LOSS
-    }
