+++ app/src/main/java/de/fraunhofer/fokus/OpenMobileNetworkToolkit/Metric/METRIC_TYPE.java
@@ -1,0 +1,9 @@
+package de.fraunhofer.fokus.OpenMobileNetworkToolkit.Metric;
+
+public enum METRIC_TYPE {
+        THROUGHPUT,
+        RTT,
+        JITTER,
+        PACKET_LOSS,
+        PING_RTT,
+    }
