+++ app/src/main/java/de/fraunhofer/fokus/OpenMobileNetworkToolkit/Ping/PingFragment.java
@@ -8,22 +8,14 @@
 
 package de.fraunhofer.fokus.OpenMobileNetworkToolkit.Ping;
 
-import static androidx.core.content.ContextCompat.getSystemService;
-
 import android.annotation.SuppressLint;
-import android.app.ActivityManager;
-import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-import android.content.IntentFilter;
 import android.content.SharedPreferences;
-import android.content.res.Configuration;
-import android.graphics.Color;
 import android.os.Bundle;
 import android.os.Handler;
 import android.text.Editable;
 import android.text.TextWatcher;
-import android.text.method.ScrollingMovementMethod;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -31,19 +23,16 @@ import android.view.ViewGroup;
 import android.widget.CompoundButton;
 import android.widget.EditText;
 import android.widget.LinearLayout;
-import android.widget.ScrollView;
 import android.widget.Switch;
 
-import android.widget.TextView;
 import androidx.fragment.app.Fragment;
-import androidx.work.WorkManager;
 
-import com.github.anastr.speedviewlib.TubeSpeedometer;
-import com.github.anastr.speedviewlib.components.Style;
-
+import de.fraunhofer.fokus.OpenMobileNetworkToolkit.Metric.METRIC_TYPE;
+import de.fraunhofer.fokus.OpenMobileNetworkToolkit.Metric.Metric;
+import java.beans.PropertyChangeEvent;
+import java.beans.PropertyChangeListener;
 import java.io.FileOutputStream;
 
-import de.fraunhofer.fokus.OpenMobileNetworkToolkit.LoggingService;
 import de.fraunhofer.fokus.OpenMobileNetworkToolkit.R;
 
 public class PingFragment extends Fragment {
@@ -55,9 +44,9 @@ public class PingFragment extends Fragment {
     private FileOutputStream stream;
     private EditText input;
     private Context ct;
-    private TextView pingViewer;
-    private ScrollView scrollView;
     private SharedPreferences sp;
+    private Metric rttMetric;
+
     public PingFragment() {
     }
     @SuppressLint("UnspecifiedRegisterReceiverFlag")
@@ -66,20 +55,24 @@ public class PingFragment extends Fragment {
         super.onCreate(savedInstanceState);
       }
 
-    private void setupPing(){
+    private void startPingService(){
         input.setEnabled(false);
-        pingViewer.setText("");
         Intent pingStart = new Intent(ct, PingService.class);
-
-        pingStart.putExtra("ping", true);
         ct.startService(pingStart);
+        rttMetric.resetMetric();
+        PingParser pingParser = PingParser.getInstance(null);
+        pingParser.addPropertyChangeListener(new PropertyChangeListener() {
+            @Override
+            public void propertyChange(PropertyChangeEvent evt) {
+                PingInformation pi = (PingInformation) evt.getNewValue();
+                rttMetric.update(pi.getRtt());
+            }
+        });
     }
-    private void stopPing(){
+    private void stopPingService(){
         input.setEnabled(true);
         Intent pingStart = new Intent(ct, PingService.class);
-        pingStart.putExtra("ping", false);
-        pingStart.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-        ct.startService(pingStart);
+        ct.stopService(pingStart);
     }
 
 
@@ -120,35 +113,20 @@ public class PingFragment extends Fragment {
         input.setText(sp.getString("ping_input", "-w 5 8.8.8.8"));
 
         ct = requireContext();
-        pingViewer = horizontalLL1.findViewById(R.id.ping_viewer);
-        scrollView = horizontalLL1.findViewById(R.id.ping_scrollviewer);
 
         saveTextInputToSharedPreferences(input, "ping_input");
-        aSwitch.setChecked(sp.getBoolean("ping", false));
+        aSwitch.setChecked(PingService.isRunning());
 
         aSwitch.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
             @Override
             public void onCheckedChanged(CompoundButton compoundButton, boolean b) {
                 Log.d(TAG, "onCheckedChanged: "+b);
-                if(b) setupPing();
-                else stopPing();
+                if(b) startPingService();
+                else stopPingService();
             }
         });
-        BroadcastReceiver receiver = new BroadcastReceiver() {
-            @Override
-            public void onReceive(Context context, Intent intent) {
-                double rtt = intent.getExtras().getDouble("ping_rtt");
-                boolean ping_running = intent.getExtras().getBoolean("ping_running");
-                handleInput(ping_running);
-                String pingLine = intent.getExtras().getString("ping_line");
-                pingViewer.append(pingLine+"\n");
-                scrollView.fullScroll(View.FOCUS_DOWN);
-            }
-        };
-        requireActivity().registerReceiver(receiver, new IntentFilter("ping"), Context.RECEIVER_EXPORTED);
-
-
-        pingViewer.setMovementMethod(new ScrollingMovementMethod());
+        rttMetric = new Metric(METRIC_TYPE.PING_RTT, ct);
+        horizontalLL1.addView(rttMetric.createOneDirection("RTT"));
         return v;
     }
 }
