+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,4 +1,12 @@
-#Thu Jun 13 12:54:04 CEST 2024
+#
+#  SPDX-FileCopyrightText: 2023 Peter Hasse <peter.hasse@fokus.fraunhofer.de>
+#  SPDX-FileCopyrightText: 2023 Johann Hackler <johann.hackler@fokus.fraunhofer.de>
+#  SPDX-FileCopyrightText: 2023 Fraunhofer FOKUS
+#
+#  SPDX-License-Identifier: BSD-3-Clause-Clear
+#
+
+#Thu Oct 12 12:17:46 CEST 2023
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 distributionUrl=https\://services.gradle.org/distributions/gradle-8.8-bin.zip
